USB: add devpath sysfs attribute
This is not exported from the usb core, yet we rely on it to create
paths to interfaces for this device in sysfs.  Export it to make
userspace tools have an easier time to figure things out.

Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
(cherry picked from commit 9af23624ae2c7978313b46e58fdc4ca5d8b799f5)