altremote: Allow to specify flash data width in Kconfig
The value by which the flash address offset of the FPGA configuration is
shifted (2 or 3 bits) can now be controlled by setting either
ALTERA_REMOTE_UPDATE_FLASH_WIDTH_8 or
ALTERA_REMOTE_UPDATE_FLASH_WIDTH_16 in the kernel configuration instead
of it being hard-coded to 3 in the driver.

Thus using this driver for FPGA images stored in EPCS flash (8 bits data
width) should now also be supported.

Signed-off-by: Tobias Klauser <klto@zhaw.ch>