altremote: Correct boot address shown in status for application mode
In application mode, the boot address read from the "current status"
set of registers seems to be bogus, but the boot address read from the
"past status 2" set of registers seems to match what we're looking for,
although this fact isn't mentioned in the Altera Remote Update user
guide.  Read the boot address from the "current status" or the "past
status 2" set of registers depending on whether it's in factory mode or
application mode, respectively.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>